+++ src/core/analytics.js
@@ -59,6 +59,7 @@ import {
 import { getIntegrationsCDNPath } from '../utils/cdnPaths';
 import { ErrorReportingService } from '../features/core/metrics/errorReporting/ErrorReportingService';
 import { getUserAgentClientHint } from '../utils/clientHint';
+import { DeviceModeTransformations } from '../features/core/deviceModeTransformation/transformationHandler';
 
 /**
  * class responsible for handling core
@@ -102,6 +103,7 @@ class Analytics {
     this.lockIntegrationsVersion = false;
     this.errorReporting = new ErrorReportingService(logger);
     this.deniedConsentIds = [];
+    this.transformationHandler = DeviceModeTransformations;
   }
 
   /**
@@ -230,6 +232,8 @@ class Analytics {
       // Initialize event repository
       this.eventRepository.initialize(this.writeKey, this.serverUrl, this.options);
       this.loaded = true;
+      // Initialize transformation handler once we determine the dataPlaneUrl
+      this.transformationHandler.init(this.writeKey, this.serverUrl, this.storage.getAuthToken());
 
       // Execute onLoaded callback if provided in load options
       if (this.options && typeof this.options.onLoaded === 'function') {
@@ -248,6 +252,10 @@ class Analytics {
           this.clientIntegrations.push({
             name: destination.destinationDefinition.name,
             config: destination.config,
+            destinationInfo: {
+              areTransformationsConnected: destination.areTransformationsConnected || false,
+              destinationId: destination.id,
+            },
           });
         }
       }, this);
@@ -310,7 +318,7 @@ class Analytics {
               const msg = `[Analytics] processResponse :: trying to initialize integration name:: ${pluginName}`;
               // logger.debug(msg);
               this.errorReporting.leaveBreadcrumb(msg);
-              intgInstance = new intMod[modName](intg.config, self);
+              intgInstance = new intMod[modName](intg.config, self, intg.destinationInfo);
               intgInstance.init();
 
               // logger.debug(pluginName, " initializing destination");
@@ -348,6 +356,120 @@ class Analytics {
       });
     } catch (error) {
       handleError(error);
+    }
+  }
+
+  /**
+   * A function to send single event to a destination
+   * @param {instance} destination
+   * @param {Object} rudderElement
+   * @param {String} methodName
+   */
+  sendDataToDestination(destination, rudderElement, methodName) {
+    try {
+      if (destination[methodName]) {
+        const clonedRudderElement = R.clone(rudderElement);
+        destination[methodName](clonedRudderElement);
+      }
+    } catch (err) {
+      const message = `[sendToNative]:: [Destination: ${destination.name}]:: `;
+      handleError(err, message);
+    }
+  }
+
+  sendTransformedDataToDestination(destWithTransformation, rudderElement, methodName) {
+    try {
+      // convert integrations object to server identified names, kind of hack now!
+      transformToServerNames(rudderElement.message.integrations);
+
+      // Process Transformation
+      this.transformationHandler.enqueue(
+        rudderElement,
+        ({ transformedPayload, transformationServerAccess }) => {
+          if (transformedPayload) {
+            destWithTransformation.forEach((intg) => {
+              try {
+                let transformedEvents = [];
+                if (transformationServerAccess) {
+                  // filter the transformed event for that destination
+                  const destTransformedResult = transformedPayload.find(
+                    (e) => e.id === intg.destinationId,
+                  );
+                  if (!destTransformedResult) {
+                    logger.error(
+                      `[DMT]::Transformed data for destination "${intg.name}" was not sent from the server`,
+                    );
+                    return;
+                  }
+
+                  destTransformedResult?.payload.forEach((tEvent) => {
+                    if (tEvent.status === '200') {
+                      transformedEvents.push(tEvent);
+                    } else {
+                      logger.error(
+                        `[DMT]::Event transformation unsuccessful for destination "${intg.name}". Dropping the event. Status: "${tEvent.status}". Error Message: "${tEvent.error}"`,
+                      );
+                    }
+                  });
+                } else {
+                  transformedEvents = transformedPayload;
+                }
+                // send transformed event to destination
+                transformedEvents?.forEach((tEvent) => {
+                  if (tEvent.event) {
+                    this.sendDataToDestination(intg, { message: tEvent.event }, methodName);
+                  }
+                });
+              } catch (e) {
+                if (e instanceof Error) {
+                  e.message = `[DMT]::[Destination:${intg.name}]:: ${e.message}`;
+                }
+                handleError(e);
+              }
+            });
+          }
+        },
+      );
+    } catch (e) {
+      if (e instanceof Error) {
+        e.message = `[DMT]:: ${e.message}`;
+      }
+      handleError(e);
+    }
+  }
+
+  /**
+   * A function to process and send events to device mode destinations
+   * @param {instance} destinations
+   * @param {Object} rudderElement
+   * @param {String} methodName
+   */
+  processAndSendEventsToDeviceMode(destinations, rudderElement, methodName) {
+    const destWithoutTransformation = [];
+    const destWithTransformation = [];
+
+    // Depending on transformation is connected or not
+    // create two sets of destinations
+    destinations.forEach((intg) => {
+      const sendEvent = !this.IsEventBlackListed(rudderElement.message.event, intg.name);
+
+      // Block the event if it is blacklisted for the device-mode destination
+      if (sendEvent) {
+        if (intg.areTransformationsConnected) {
+          destWithTransformation.push(intg);
+        } else {
+          destWithoutTransformation.push(intg);
+        }
+      }
+    });
+    // loop through destinations that doesn't have
+    // transformation connected with it and send events
+    destWithoutTransformation.forEach((intg) => {
+      this.sendDataToDestination(intg, rudderElement, methodName);
+    });
+
+    if (destWithTransformation.length > 0) {
+      this.sendTransformedDataToDestination(destWithTransformation, rudderElement, methodName);
     }
   }
 
@@ -405,28 +527,11 @@ class Analytics {
       );
 
       // send to all integrations now from the 'toBeProcessedByIntegrationArray' replay queue
-      for (const successfulLoadedIntersectClientSuppliedIntegration of successfulLoadedIntersectClientSuppliedIntegrations) {
-        try {
-          if (
-            (!successfulLoadedIntersectClientSuppliedIntegration.isFailed ||
-              !successfulLoadedIntersectClientSuppliedIntegration.isFailed()) &&
-            successfulLoadedIntersectClientSuppliedIntegration[methodName]
-          ) {
-            const sendEvent = !object.IsEventBlackListed(
-              event[0].message.event,
-              successfulLoadedIntersectClientSuppliedIntegration.name,
-            );
-
-            // Block the event if it is blacklisted for the device-mode destination
-            if (sendEvent) {
-              const clonedBufferEvent = R.clone(event);
-              successfulLoadedIntersectClientSuppliedIntegration[methodName](...clonedBufferEvent);
-            }
-          }
-        } catch (error) {
-          handleError(error);
-        }
-      }
+      this.processAndSendEventsToDeviceMode(
+        successfulLoadedIntersectClientSuppliedIntegrations,
+        event[0],
+        methodName,
+      );
     });
     object.toBeProcessedByIntegrationArray = [];
   }
@@ -774,22 +879,11 @@ class Analytics {
         );
 
         // try to first send to all integrations, if list populated from BE
-        successfulLoadedIntersectClientSuppliedIntegrations.forEach((obj) => {
-          try {
-            if ((!obj.isFailed || !obj.isFailed()) && obj[type]) {
-              const sendEvent = !this.IsEventBlackListed(rudderElement.message.event, obj.name);
-
-              // Block the event if it is blacklisted for the device-mode destination
-              if (sendEvent) {
-                const clonedRudderElement = R.clone(rudderElement);
-                obj[type](clonedRudderElement);
-              }
-            }
-          } catch (err) {
-            const message = `[sendToNative]:: [Destination: ${obj.name}]:: `;
-            handleError(err, message);
-          }
-        });
+        this.processAndSendEventsToDeviceMode(
+          successfulLoadedIntersectClientSuppliedIntegrations,
+          rudderElement,
+          type,
+        );
       }
 
       // convert integrations object to server identified names, kind of hack now!
@@ -1112,6 +1206,29 @@ class Analytics {
     }
   }
 
+  arePolyfillsRequired(options) {
+    // check if the below features are available in the browser or not
+    // If not present dynamically load from the polyfill cdn, unless
+    // the options are configured not to.
+    const polyfillIfRequired =
+      options && typeof options.polyfillIfRequired === 'boolean'
+        ? options.polyfillIfRequired
+        : true;
+    return (
+      polyfillIfRequired &&
+      (!String.prototype.endsWith ||
+        !String.prototype.startsWith ||
+        !String.prototype.includes ||
+        !Array.prototype.find ||
+        !Array.prototype.includes ||
+        !Promise ||
+        !Object.entries ||
+        !Object.values ||
+        !String.prototype.replaceAll ||
+        !this.isDatasetAvailable())
+    );
+  }
+
   /**
    * Call control pane to get client configs
    *
@@ -1124,26 +1241,7 @@ class Analytics {
 
     // clone options
     const clonedOptions = R.clone(options);
-    // check if the below features are available in the browser or not
-    // If not present dynamically load from the polyfill cdn, unless
-    // the options are configured not to.
-    const polyfillIfRequired =
-      clonedOptions && typeof clonedOptions.polyfillIfRequired === 'boolean'
-        ? clonedOptions.polyfillIfRequired
-        : true;
-    if (
-      polyfillIfRequired &&
-      (!String.prototype.endsWith ||
-        !String.prototype.startsWith ||
-        !String.prototype.includes ||
-        !Array.prototype.find ||
-        !Array.prototype.includes ||
-        !Promise ||
-        !Object.entries ||
-        !Object.values ||
-        !String.prototype.replaceAll ||
-        !this.isDatasetAvailable())
-    ) {
+    if (this.arePolyfillsRequired(clonedOptions)) {
       const id = 'polyfill';
       ScriptLoader(id, POLYFILL_URL, { skipDatasetAttributes: true });
       const self = this;
@@ -1250,6 +1348,15 @@ class Analytics {
    */
   endSession() {
     this.uSession.end();
+  }
+
+  setAuthToken(token) {
+    if (typeof token !== 'string') {
+      logger.error('Provided input should be in string format');
+      return;
+    }
+    this.storage.setAuthToken(token);
+    this.transformationHandler.setAuthToken(token);
   }
 }
 
@@ -1378,6 +1485,7 @@ const getGroupId = instance.getGroupId.bind(instance);
 const getGroupTraits = instance.getGroupTraits.bind(instance);
 const startSession = instance.startSession.bind(instance);
 const endSession = instance.endSession.bind(instance);
+const setAuthToken = instance.setAuthToken.bind(instance);
 
 export {
   initialized,
@@ -1398,4 +1506,5 @@ export {
   getGroupTraits,
   startSession,
   endSession,
+  setAuthToken,
 };