+++ src/integrations/GA/index.test.js
@@ -7,12 +7,14 @@ afterAll(() => {
   jest.restoreAllMocks();
 });
 
+const destinationInfo = { areTransformationsConnected: false, destinationId: 'sample-destination-id' };
+
 GA.prototype.loadScript = jest.fn();
 
 describe('GA init tests', () => {
   let googleAnalytics;
   beforeEach(() => {
-    googleAnalytics = new GA({ trackingID: 'UA-143161493-8' }, { loadIntegration: true });
+    googleAnalytics = new GA({ trackingID: 'UA-143161493-8' }, { loadIntegration: true }, destinationInfo);
     googleAnalytics.init();
   });
 
@@ -61,6 +63,7 @@ describe('GA init tests', () => {
           ],
         },
         { loadIntegration: true },
+        destinationInfo
       );
       googleAnalytics.init();
       window.ga = jest.fn();
@@ -110,6 +113,7 @@ describe('GA init tests', () => {
           contentGroupings: [],
         },
         { loadIntegration: true },
+        destinationInfo
       );
       googleAnalytics.init();
       window.ga = jest.fn();