+++ CHANGELOG.md
@@ -2,6 +2,50 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.32.0](https://github.com/rudderlabs/rudder-sdk-js/compare/v2.31.1...v2.32.0) (2023-04-24)
+
+
+### Features
+
+* **pinterest:** custom event ([#1027](https://github.com/rudderlabs/rudder-sdk-js/issues/1027)) ([636b0b8](https://github.com/rudderlabs/rudder-sdk-js/commit/636b0b8a3ee7137ca9b637d054a7a88c116c8ae9))
+
+### [2.31.1](https://github.com/rudderlabs/rudder-sdk-js/compare/v2.31.0...v2.31.1) (2023-04-20)
+
+
+### Bug Fixes
+
+* integrations not being compatible with older versions of analytics ([36fe09b](https://github.com/rudderlabs/rudder-sdk-js/commit/36fe09bd804a3aafc38b214308ceedce54d027a3))
+* remove duplicate conditional check in VWO integration ([99a5e4b](https://github.com/rudderlabs/rudder-sdk-js/commit/99a5e4b0c22a0580304a551dc6704cd0c3b2a614))
+
+## [2.31.0](https://github.com/rudderlabs/rudder-sdk-js/compare/v2.30.0...v2.31.0) (2023-04-20)
+
+
+### Features
+
+* device mode destination ([#1014](https://github.com/rudderlabs/rudder-sdk-js/issues/1014)) ([cc96e22](https://github.com/rudderlabs/rudder-sdk-js/commit/cc96e22c06e9ff96fc547d22d3f1ca592122593f))
+* **GA4:** override sessionId with rudder sessionId ([#988](https://github.com/rudderlabs/rudder-sdk-js/issues/988)) ([a62717a](https://github.com/rudderlabs/rudder-sdk-js/commit/a62717a32329c889b9b33ac87cbad0e30f0feb99))
+
+
+### Bug Fixes
+
+* bugsnag typo in rollup config ([#1010](https://github.com/rudderlabs/rudder-sdk-js/issues/1010)) ([f085feb](https://github.com/rudderlabs/rudder-sdk-js/commit/f085febde59ea74d64927a112526e7e51d5fa2b4))
+
+## [2.30.0](https://github.com/rudderlabs/rudder-sdk-js/compare/v2.29.0...v2.30.0) (2023-04-17)
+
+
+### Features
+
+* bingads Upgraded to latest ([#1003](https://github.com/rudderlabs/rudder-sdk-js/issues/1003)) ([24976a4](https://github.com/rudderlabs/rudder-sdk-js/commit/24976a4fa5b507eba3444b27f2357c71f25df6ad))
+
+
+### Bug Fixes
+
+* bugsnag error reporting ([#856](https://github.com/rudderlabs/rudder-sdk-js/issues/856)) ([b35bd2b](https://github.com/rudderlabs/rudder-sdk-js/commit/b35bd2bca1dc372b863fb7afe028061507d2b64d))
+* dom-exception error in incognito for storejs lib ([#991](https://github.com/rudderlabs/rudder-sdk-js/issues/991)) ([66bb84f](https://github.com/rudderlabs/rudder-sdk-js/commit/66bb84fb27a1a61d5bfba47015c363cb7dcdb923))
+* **fb_pixel:** add value in mapped event ([#1005](https://github.com/rudderlabs/rudder-sdk-js/issues/1005)) ([945e3dd](https://github.com/rudderlabs/rudder-sdk-js/commit/945e3dd49742585de1b0e70337156eefe6fe3e84))
+* **GA4:** add send_to parameter support by default ([#993](https://github.com/rudderlabs/rudder-sdk-js/issues/993)) ([05cdaba](https://github.com/rudderlabs/rudder-sdk-js/commit/05cdabac31a0f246e5343b0622142cdefee94ec2))
+* upgrade vulnerable dependencies ([#995](https://github.com/rudderlabs/rudder-sdk-js/issues/995)) ([8e6d8ff](https://github.com/rudderlabs/rudder-sdk-js/commit/8e6d8ff9a0712f1fbf3672829de5b9264266f716))
+
 ## [2.29.0](https://github.com/rudderlabs/rudder-sdk-js/compare/v2.28.0...v2.29.0) (2023-04-03)